+++ opyoid/bindings/provider_binding/provider_binding.py
@@ -6,7 +6,7 @@ from opyoid.bindings.binding import Binding
 from opyoid.exceptions import BindingError
 from opyoid.provider import Provider
 from opyoid.scopes import Scope, SingletonScope
-from opyoid.utils import InjectedT, get_class_full_name, get_function_full_name
+from opyoid.utils import get_class_full_name, get_function_full_name, InjectedT
 
 
 @attr.s(auto_attribs=True, frozen=True, repr=False)