+++ opyoid/frozen_target.py
@@ -2,7 +2,7 @@ from typing import Generic, Optional, Type
 
 import attr
 
-from opyoid.utils import InjectedT, get_class_full_name
+from opyoid.utils import get_class_full_name, InjectedT
 
 
 @attr.s(auto_attribs=True, frozen=True, repr=False)